requirements.txt
pyrogram==2.0.106
TgCrypto
motor
aiohttp
aiofiles
python-dotenv
pymongo
